feature: convert irc/discord modules to ArgumentParser's
